requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
requests>=2.31.0
httpx>=0.27.0
numpy>=1.26.0
redis>=5.0.0
python-multipart>=0.0.9
apscheduler>=3.10.0
pyyaml>=6.0
